content_service/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
mysqlclient>=2.1.0
Pillow>=10.0.0
djangorestframework~=3.15.2
django-cors-headers